+++ tests/testthat/test-tetra_example.R
@@ -198,13 +198,15 @@ test_that("upload files",{
   expect_equal(sum(data.prob$homoprob$probability), 14900, tolerance = 0.001)
 
   input.haplo <- c("Trait:SG06_LG:2_Pos:77_homolog:P1.1")
-  p <- select_haplo(input.haplo,
+  p1.list <- select_haplo(input.haplo,
                     viewpoly_obj$qtl$probs,
                     viewpoly_obj$qtl$selected_mks,
                     effects.data = p)
-  expect_equal(sum(p[[1]]$data$probability), 507.9996, tolerance = 0.0001)
-  expect_equal(sum(p[[2]]$data$probability), 508.001, tolerance = 0.0001)
-  expect_equal(sum(p[[3]]$data$probability), 508.0009, tolerance = 0.0001)
+  
+  p1 <- p1.list[[1]]
+  expect_equal(sum(p1[[1]]$data$probability), 507.9996, tolerance = 0.0001)
+  expect_equal(sum(p1[[2]]$data$probability), 508.001, tolerance = 0.0001)
+  expect_equal(sum(p1[[3]]$data$probability), 508.0009, tolerance = 0.0001)
 
   # VIEWgenome tests
   p <- plot_cm_mb(viewpoly_obj$map, 1, 1,50)